+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 
+- concord-server: new `agent-workers-available` metric, shows how
+many Agent worker slots are available;
 - http-tasks: support for `multipart/form-data` requests;
 - concord-agent: support for running processes with custom JVM
 parameters.